@@ -23,7 +23,7 @@
   },
   "license": "MIT",
   "dependencies": {
-    "@docusaurus/logger": "^2.0.0",
+    "@docusaurus/logger": "^3.0.0",
     "commander": "^5.1.0",
     "fs-extra": "^10.0.0",
     "lodash": "^4.17.20",
@@ -34,7 +34,7 @@
     "tslib": "^2.3.1"
   },
   "engines": {
-    "node": ">=14"
+    "node": ">=18"
   },
   "devDependencies": {
     "@types/prompts": "^2.0.9",
