@@ -28,19 +28,19 @@
     "watch": "tsc --watch"
   },
   "devDependencies": {
-    "@docusaurus/types": "^2.0.0"
+    "@docusaurus/types": "^3.0.0"
   },
   "dependencies": {
-    "@docusaurus/preset-classic": "^2.0.0",
+    "@docusaurus/preset-classic": "^3.0.0",
     "docusaurus-plugin-openapi": "^0.6.4",
     "docusaurus-plugin-proxy": "^0.6.3",
     "docusaurus-theme-openapi": "^0.6.4"
   },
   "peerDependencies": {
-    "react": "^16.8.4 || ^17.0.0",
-    "react-dom": "^16.8.4 || ^17.0.0"
+    "react": "^18.0.0",
+    "react-dom": "^18.0.0"
   },
   "engines": {
-    "node": ">=14"
+    "node": ">=18"
   }
 }
