@@ -24,8 +24,8 @@
     "format:lib-next": "prettier --config ../../.prettierrc.json --write \"lib-next/**/*.{js,ts,jsx,tsc}\""
   },
   "devDependencies": {
-    "@docusaurus/module-type-aliases": "^2.0.0",
-    "@docusaurus/types": "^2.0.0",
+    "@docusaurus/module-type-aliases": "^3.0.0",
+    "@docusaurus/types": "^3.0.0",
     "@types/concurrently": "^6.3.0",
     "@types/crypto-js": "^4.1.0",
     "@types/fs-extra": "^9.0.13",
@@ -36,12 +36,12 @@
     "concurrently": "^5.2.0"
   },
   "dependencies": {
-    "@docusaurus/theme-common": "^2.0.0",
-    "@mdx-js/react": "^1.6.22",
+    "@docusaurus/theme-common": "^3.0.0",
+    "@mdx-js/react": "^3.0.0",
     "@monaco-editor/react": "^4.3.1",
     "@reduxjs/toolkit": "^1.7.1",
     "buffer": "^6.0.3",
-    "clsx": "^1.1.1",
+    "clsx": "^1.2.1",
     "crypto-js": "^4.1.1",
     "docusaurus-plugin-openapi": "^0.6.4",
     "immer": "^9.0.7",
@@ -49,18 +49,18 @@
     "monaco-editor": "^0.31.1",
     "postman-code-generators": "^1.0.0",
     "postman-collection": "^4.1.0",
-    "prism-react-renderer": "^1.2.1",
+    "prism-react-renderer": "^2.1.0",
     "process": "^0.11.10",
     "react-magic-dropzone": "^1.0.1",
     "react-redux": "^7.2.0",
     "redux-devtools-extension": "^2.13.8",
-    "webpack": "^5.73.0"
+    "webpack": "^5.88.1"
   },
   "peerDependencies": {
-    "react": "^16.8.4 || ^17.0.0",
-    "react-dom": "^16.8.4 || ^17.0.0"
+    "react": "^18.0.0",
+    "react-dom": "^18.0.0"
   },
   "engines": {
-    "node": ">=14"
+    "node": ">=18"
   }
 }
