@@ -28,4 +28,4 @@ const sidebars = {
    */
 };
 
-module.exports = sidebars;
+export default sidebars;
