@@ -85,7 +85,7 @@ function createItems(openapiData: OpenApiObject): ApiMetadata[] {
     const infoPage: PartialPage<InfoPageMetadata> = {
       type: "info",
       id: "introduction",
-      unversionedId: "introduction",
+      // unversionedId: "introduction",
       title: "Introduction",
       description: openapiData.info.description,
       slug: "/introduction",
@@ -144,7 +144,7 @@ function createItems(openapiData: OpenApiObject): ApiMetadata[] {
       const apiPage: PartialPage<ApiPageMetadata> = {
         type: "api",
         id: baseId,
-        unversionedId: baseId,
+        // unversionedId: baseId,
         title: title,
         description: description ?? "",
         slug: "/" + baseId,
@@ -298,7 +298,7 @@ export async function processOpenapiFiles(
     }
 
     items[i].id = id;
-    items[i].unversionedId = id;
+    // items[i].unversionedId = id;
     items[i].slug = "/" + id;
   }
 
