@@ -49,7 +49,7 @@ export interface ApiMetadataBase {
   next?: ApiNavLink;
   //
   id: string; // TODO legacy versioned id => try to remove
-  unversionedId: string; // TODO new unversioned id => try to rename to "id"
+  // unversionedId: string; // TODO new unversioned id => try to rename to "id"
   title: string;
   description: string;
   source: string; // @site aliased source => "@site/docs/folder/subFolder/subSubFolder/myDoc.md"
